Basics of the report viewer finished
To do:

Finish the read only user stuff
Figure out what the fuck is going on with struts 2.17 and why the parameter stuff doesnt work in 2.16

and

all the admin screen stuff

but

getting there :)

+++ RR3/src/main/java/binky/reportrunner/ui/actions/job/DeleteJob.java
@@ -20,7 +20,7 @@ public class DeleteJob extends StandardRunnerAction {
 		if (groupName != null && !groupName.isEmpty()
 				&& (jobName != null && !jobName.isEmpty())) {
 			// security check
-			if (doesUserHaveGroup(groupName)) {
+			if (doesUserHaveGroup(groupName) && isUserReadOnly()) {
 				jobService.deleteJob(jobName, groupName);
 			} else {
 				SecurityException se = new SecurityException("Group "
